[VM, Content] Preliminary TS1 stuff, improvements.
- TS1 world content system (walls and floors still read from TSO)
- ARRY, CFP, BCF, BCP, CARR (with VM implementation), WALm, FLRm, OBJT,
OBJM (VERY PRELIMINARY, using an analytic approach rather than reading
the data structure accurately)
- TS1 Hybrid mode, with loading HouseXX.iff. (lot zoning type and
startup lot are currently hardcoded)
- TS1 Lot activator
- Set To Next category now does the right thing
- Preliminary speed multiplier support
- Generalize Lightmaps for lots of varying sizes
- Sort of really hacky outfit dual typing, to be cleaned up.
- Minor optimisations to animations and skill headlines (more to come)
- Crashfix for money over head.

+++ TSOClient/tso.simantics/primitives/VMSetBalloonHeadline.cs
@@ -115,7 +115,7 @@ namespace FSO.SimAntics.Primitives
             using (var io = IoBuffer.FromBytes(bytes, ByteOrder.LITTLE_ENDIAN)){
                 Flags2 = io.ReadUInt16();
                 Index = (sbyte)io.ReadByte();
-                Group = (VMSetBalloonHeadlineOperandGroup)(sbyte)io.ReadByte();
+                Group = (VMSetBalloonHeadlineOperandGroup)io.ReadByte();
                 Duration = io.ReadInt16();
                 Type = io.ReadByte();
                 Flags = io.ReadByte();
